requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
requests==2.31.0
xmltodict==0.13.0
